move everything into POVS component
